bsp(M5Stack Tab5): Update LCD and touch for new board version + added BMI270 sensor support
