+++ bsp/m5stack_tab5/idf_component.yml
@@ -1,4 +1,5 @@
-version: "1.1.0"
+
+version: "1.2.0"
 description: Board Support Package (BSP) for m5stack_tab5
 url: https://github.com/espressif/esp-bsp/tree/master/bsp/m5stack_tab5
 
@@ -13,7 +14,7 @@ dependencies:
 
   esp_codec_dev:
     version: "~1.5"
-    public:  true
+    public: true
 
   espressif/esp_io_expander:
     version: "*"
@@ -22,34 +23,49 @@ dependencies:
 
   espressif/esp_io_expander_pi4ioe5v6408:
     version: "*"
-    public:  false
+    public: true
     override_path: "../../components/io_expander/esp_io_expander_pi4ioe5v6408"
 
   espressif/esp_lvgl_port:
     version: "^2"
-    public:  true
+    public: true
     override_path: "../../components/esp_lvgl_port"
+
+  espressif/esp_lcd_st7123:
+    version: "*"
+    public: false
 
   espressif/esp_lcd_ili9881c:
     version: "*"
-    public:  false
+    public: false
+    override_path: "../../components/lcd/esp_lcd_ili9881c"
+
+  espressif/esp_lcd_touch_st7123:
+    version: "*"
+    public: false
 
   espressif/esp_lcd_touch_gt911:
     version: "*"
-    public:  false
-    override_path: "../../components/lcd_touch/esp_lcd_touch_gt911"
+    public: false
+    # Here cannot be override_path, because we are using two touch comonents.
+    # The ST7123 is not owned by BSP and loading another version of esp_lcd_touch
 
   esp_video:
     version: "~2.0"
-    public:  true
+    public: true
 
   usb:
     version: "^1"
-    public:  true
+    public: true
+
+  espressif/bmi270:
+    version: "*"
+    public: true
+    override_path: "../../components/sensors/bmi270"
 
 examples:
   - path: ../../examples/display
-  - path: ../../examples/display_rotation
+  - path: ../../examples/display_sensors
   - path: ../../examples/display_audio_photo
   - path: ../../examples/display_lvgl_demos
   - path: ../../examples/display_camera_video
