+++ bsp/m5stack_tab5/src/bsp_display.c
@@ -1,0 +1,553 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "esp_err.h"
+#include "esp_log.h"
+#include "esp_check.h"
+#include "bsp_err_check.h"
+#include "driver/ledc.h"
+#include "esp_lcd_panel_io.h"
+#include "esp_lcd_panel_vendor.h"
+#include "esp_lcd_panel_ops.h"
+#include "esp_lcd_mipi_dsi.h"
+#include "esp_ldo_regulator.h"
+#include "sd_pwr_ctrl_by_on_chip_ldo.h"
+#include "esp_lcd_ili9881c.h"
+#include "esp_lcd_st7123.h"
+#include "disp_init_data.h"
+#include "disp_init_data_1.h"
+#include "esp_lcd_touch_gt911.h"
+#include "esp_lcd_touch_st7123.h"
+#include "bsp/display.h"
+#include "bsp/touch.h"
+
+#include "bsp/m5stack_tab5.h"
+
+static const char *TAG = "m5stack_tab5";
+
+
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static lv_display_t *disp;
+static lv_indev_t *disp_indev_touch = NULL;
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static bsp_lcd_handles_t disp_handles;
+static esp_ldo_channel_handle_t disp_phy_pwr_chan = NULL;
+static esp_lcd_touch_handle_t tp;   // LCD touch handle
+
+// Bit number used to represent command and parameter
+#define LCD_CMD_BITS           8
+#define LCD_PARAM_BITS         8
+#define LCD_LEDC_CH            CONFIG_BSP_DISPLAY_BRIGHTNESS_LEDC_CH
+
+esp_err_t bsp_display_brightness_init(void)
+{
+    // Backlight on normal GPIO: Setup LEDC peripheral for PWM
+    const ledc_channel_config_t LCD_backlight_channel = {
+        .gpio_num = BSP_LCD_BACKLIGHT,
+        .speed_mode = LEDC_LOW_SPEED_MODE,
+        .channel = LCD_LEDC_CH,
+        .intr_type = LEDC_INTR_DISABLE,
+        .timer_sel = LEDC_TIMER_0,
+        .duty = 0,
+        .hpoint = 0
+    };
+    const ledc_timer_config_t LCD_backlight_timer = {
+        .speed_mode = LEDC_LOW_SPEED_MODE,
+        .duty_resolution = LEDC_TIMER_10_BIT,
+        .timer_num = LEDC_TIMER_0,
+        .freq_hz = 5000,
+        .clk_cfg = LEDC_AUTO_CLK
+    };
+
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_channel_config(&LCD_backlight_channel));
+
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_brightness_deinit(void)
+{
+    const ledc_timer_config_t LCD_backlight_timer = {
+        .speed_mode = LEDC_LOW_SPEED_MODE,
+        .timer_num = LEDC_TIMER_0,
+        .deconfigure = 1
+    };
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_pause(LEDC_LOW_SPEED_MODE, LEDC_TIMER_0));
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_brightness_set(int brightness_percent)
+{
+    if (brightness_percent > 100) {
+        brightness_percent = 100;
+    }
+    if (brightness_percent < 0) {
+        brightness_percent = 0;
+    }
+
+    ESP_LOGI(TAG, "Setting LCD backlight: %d%%", brightness_percent);
+    uint32_t duty_cycle = (1023 * brightness_percent) / 100; // LEDC resolution set to 10bits, thus: 100% = 1023
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_set_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH, duty_cycle));
+    BSP_ERROR_CHECK_RETURN_ERR(ledc_update_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH));
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_backlight_off(void)
+{
+    return bsp_display_brightness_set(0);
+}
+
+esp_err_t bsp_display_backlight_on(void)
+{
+    return bsp_display_brightness_set(100);
+}
+static esp_err_t bsp_enable_dsi_phy_power(void)
+{
+#if BSP_MIPI_DSI_PHY_PWR_LDO_CHAN > 0
+    // Turn on the power for MIPI DSI PHY, so it can go from "No Power" state to "Shutdown" state
+    esp_ldo_channel_config_t ldo_cfg = {
+        .chan_id = BSP_MIPI_DSI_PHY_PWR_LDO_CHAN,
+        .voltage_mv = BSP_MIPI_DSI_PHY_PWR_LDO_VOLTAGE_MV,
+    };
+    ESP_RETURN_ON_ERROR(esp_ldo_acquire_channel(&ldo_cfg, &disp_phy_pwr_chan), TAG, "Acquire LDO channel for DPHY failed");
+    ESP_LOGI(TAG, "MIPI DSI PHY Powered on");
+#endif // BSP_MIPI_DSI_PHY_PWR_LDO_CHAN > 0
+
+    return ESP_OK;
+}
+
+/*
+ * !!! This function was manually added to generated code !!!
+ * It selects board version by LCD touch I2C driver probing
+ * Returns
+ * 1: LCD ILI9881C, Touch GT911
+ * 2: LCD ST7123, Touch ST7123
+ */
+static int bsp_get_board_version(void)
+{
+    static int board_ver = 0;
+    if (board_ver > 0) {
+        return board_ver;
+    }
+
+    /* Enable Feature */
+    BSP_ERROR_CHECK_RETURN_NULL(bsp_feature_enable(BSP_FEATURE_TOUCH, true));
+
+    vTaskDelay(pdMS_TO_TICKS(500));
+
+    if (i2c_master_probe(bsp_i2c_get_handle(), ESP_LCD_TOUCH_IO_I2C_ST7123_ADDRESS, 100) == ESP_OK) {
+        ESP_LOGI(TAG, "Discovered board version 2 (LCD ST7123, Touch ST7123)");
+        board_ver = 2;
+    } else if (i2c_master_probe(bsp_i2c_get_handle(), ESP_LCD_TOUCH_IO_I2C_GT911_ADDRESS_BACKUP, 100) == ESP_OK) {
+        ESP_LOGI(TAG, "Discovered board version 1 (LCD ILI9881C, Touch GT911)");
+        board_ver = 1;
+    } else {
+        ESP_LOGE(TAG, "Unsupported board version!");
+        assert(NULL);
+    }
+
+    return board_ver;
+}
+
+esp_err_t bsp_display_new(const bsp_display_config_t *config, esp_lcd_panel_handle_t *ret_panel,
+                          esp_lcd_panel_io_handle_t *ret_io)
+{
+    esp_err_t ret = ESP_OK;
+    bsp_lcd_handles_t handles;
+    ret = bsp_display_new_with_handles(config, &handles);
+
+    *ret_panel = handles.panel;
+    *ret_io = handles.io;
+
+    return ret;
+}
+
+esp_err_t bsp_display_new_with_handles(const bsp_display_config_t *config, bsp_lcd_handles_t *ret_handles)
+{
+    esp_err_t ret = ESP_OK;
+
+    /* Enable Feature */
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_feature_enable(BSP_FEATURE_LCD, true));
+    ESP_RETURN_ON_ERROR(bsp_display_brightness_init(), TAG, "Brightness init failed");
+    ESP_RETURN_ON_ERROR(bsp_enable_dsi_phy_power(), TAG, "DSI PHY power failed");
+
+    /* create MIPI DSI bus first, it will initialize the DSI PHY as well */
+    esp_lcd_dsi_bus_config_t bus_config = {
+        .bus_id = 0,
+        .num_data_lanes = BSP_LCD_MIPI_DSI_LANE_NUM,
+        .phy_clk_src = config->dsi_bus.phy_clk_src,
+        .lane_bit_rate_mbps = config->dsi_bus.lane_bit_rate_mbps,
+    };
+    ESP_RETURN_ON_ERROR(esp_lcd_new_dsi_bus(&bus_config, &ret_handles->mipi_dsi_bus), TAG, "New DSI bus init failed");
+    disp_handles.mipi_dsi_bus = ret_handles->mipi_dsi_bus;
+
+    ESP_LOGI(TAG, "Install MIPI DSI LCD control panel");
+    // we use DBI interface to send LCD commands and parameters
+    esp_lcd_dbi_io_config_t dbi_config = {
+        .virtual_channel = 0,
+        .lcd_cmd_bits = 8,   // according to the LCD spec
+        .lcd_param_bits = 8, // according to the LCD spec
+    };
+    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_io_dbi(ret_handles->mipi_dsi_bus, &dbi_config, &ret_handles->io), err, TAG,
+                      "New panel IO failed");
+    disp_handles.io = ret_handles->io;
+
+    const esp_lcd_dpi_panel_config_t dpi_config_ili9881c = {
+        .virtual_channel    = 0,
+        .dpi_clk_src        = MIPI_DSI_DPI_CLK_SRC_DEFAULT,
+        .dpi_clock_freq_mhz = 60,
+        .in_color_format    = LCD_COLOR_FMT_RGB565,
+        .num_fbs            = CONFIG_BSP_LCD_DPI_BUFFER_NUMS,
+        .video_timing =
+        {
+            .h_size            = BSP_LCD_H_RES,
+            .v_size            = BSP_LCD_V_RES,
+            .hsync_back_porch  = 140,
+            .hsync_pulse_width = 40,
+            .hsync_front_porch = 40,
+            .vsync_back_porch  = 20,
+            .vsync_pulse_width = 4,
+            .vsync_front_porch = 20,
+        },
+#if CONFIG_BSP_LCD_USE_DMA2D && (ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(6, 0, 0))
+        .flags.use_dma2d = true,
+#endif
+    };
+    const esp_lcd_dpi_panel_config_t dpi_config_st7123 = {
+        .virtual_channel    = 0,
+        .dpi_clk_src        = MIPI_DSI_DPI_CLK_SRC_DEFAULT,
+        .dpi_clock_freq_mhz = 70,
+        .in_color_format    = LCD_COLOR_FMT_RGB565,
+        .num_fbs            = CONFIG_BSP_LCD_DPI_BUFFER_NUMS,
+        .video_timing =
+        {
+            .h_size            = BSP_LCD_H_RES,
+            .v_size            = BSP_LCD_V_RES,
+            .hsync_back_porch  = 40,
+            .hsync_pulse_width = 2,
+            .hsync_front_porch = 40,
+            .vsync_back_porch  = 8,
+            .vsync_pulse_width = 2,
+            .vsync_front_porch = 220,
+        },
+#if CONFIG_BSP_LCD_USE_DMA2D && (ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(6, 0, 0))
+        .flags.use_dma2d = true,
+#endif
+    };
+
+    const void *vendor_config = NULL;
+    const ili9881c_vendor_config_t vendor_config_ili9881c = {
+        .init_cmds      = disp_init_data_ili9881c,
+        .init_cmds_size = sizeof(disp_init_data_ili9881c) / sizeof(disp_init_data_ili9881c[0]),
+        .mipi_config = {
+            .dsi_bus = ret_handles->mipi_dsi_bus,
+            .dpi_config = &dpi_config_ili9881c,
+            .lane_num   = BSP_LCD_MIPI_DSI_LANE_NUM,
+        },
+    };
+    const st7123_vendor_config_t vendor_config_st7123 = {
+        .init_cmds      = disp_init_data_st7123,
+        .init_cmds_size = sizeof(disp_init_data_st7123) / sizeof(disp_init_data_st7123[0]),
+        .mipi_config = {
+            .dsi_bus = ret_handles->mipi_dsi_bus,
+            .dpi_config = &dpi_config_st7123,
+        },
+    };
+
+    /* Select vendor config by board version */
+    if (bsp_get_board_version() == 1) {
+        vendor_config = &vendor_config_ili9881c;
+    } else if (bsp_get_board_version() == 2) {
+        vendor_config = &vendor_config_st7123;
+    }
+
+    ESP_LOGD(TAG, "Install LCD driver");
+    const esp_lcd_panel_dev_config_t panel_config = {
+        .reset_gpio_num = BSP_LCD_RST,
+        .flags.reset_active_high = 0,
+        .rgb_ele_order = BSP_LCD_COLOR_SPACE,
+        .bits_per_pixel = BSP_LCD_BITS_PER_PIXEL,
+        .vendor_config = (void *)vendor_config,
+    };
+
+    /* Select LCD panel by board version */
+    if (bsp_get_board_version() == 1) {
+        ESP_GOTO_ON_ERROR(esp_lcd_new_panel_ili9881c(ret_handles->io, &panel_config, &ret_handles->panel),
+                          err, TAG, "New panel failed");
+    } else if (bsp_get_board_version() == 2) {
+        ESP_GOTO_ON_ERROR(esp_lcd_new_panel_st7123(ret_handles->io, &panel_config, &ret_handles->panel),
+                          err, TAG, "New panel failed");
+    }
+    disp_handles.panel = ret_handles->panel;
+#if CONFIG_BSP_LCD_USE_DMA2D && (ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(6, 0, 0))
+    ESP_GOTO_ON_ERROR(esp_lcd_dpi_panel_enable_dma2d(ret_handles->panel), err, TAG, "LCD panel enable DMA2D failed");
+#endif
+
+    esp_lcd_panel_reset(ret_handles->panel);
+    esp_lcd_panel_init(ret_handles->panel);
+    esp_lcd_panel_invert_color(ret_handles->panel, false);
+    esp_lcd_panel_mirror(ret_handles->panel, false, false);
+
+    ESP_LOGI(TAG, "Display initialized with resolution %dx%d", BSP_LCD_H_RES, BSP_LCD_V_RES);
+    return ret;
+
+err:
+    bsp_display_delete();
+    return ret;
+}
+
+void bsp_display_delete(void)
+{
+    if (disp_handles.panel) {
+        esp_lcd_panel_del(disp_handles.panel);
+        disp_handles.panel = NULL;
+    }
+    if (disp_handles.io) {
+        esp_lcd_panel_io_del(disp_handles.io);
+        disp_handles.io = NULL;
+    }
+    if (disp_handles.mipi_dsi_bus) {
+        esp_lcd_del_dsi_bus(disp_handles.mipi_dsi_bus);
+        disp_handles.mipi_dsi_bus = NULL;
+    }
+
+    if (disp_phy_pwr_chan) {
+        esp_ldo_release_channel(disp_phy_pwr_chan);
+        disp_phy_pwr_chan = NULL;
+    }
+
+    bsp_display_brightness_deinit();
+}
+
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static lv_display_t *bsp_display_lcd_init(const bsp_display_cfg_t *cfg)
+{
+    assert(cfg != NULL);
+    esp_lcd_panel_io_handle_t io_handle = NULL;
+    const bsp_display_config_t bsp_disp_cfg = {
+        .dsi_bus = {
+            .phy_clk_src = 0, // let the driver to choose the default clock source
+            .lane_bit_rate_mbps = BSP_LCD_MIPI_DSI_LANE_BITRATE_MBPS,
+        }
+    };
+    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_new(&bsp_disp_cfg, &disp_handles.panel, &io_handle));
+
+    esp_lcd_panel_disp_on_off(disp_handles.panel, true);
+
+    /* Add LCD screen */
+    ESP_LOGD(TAG, "Add LCD screen");
+    const lvgl_port_display_cfg_t disp_cfg = {
+        .io_handle = io_handle,
+        .panel_handle = disp_handles.panel,
+        .buffer_size = cfg->buffer_size,
+        .double_buffer = cfg->double_buffer,
+        .hres = BSP_LCD_H_RES,
+        .vres = BSP_LCD_V_RES,
+        .monochrome = false,
+        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
+        .rotation = {
+            .swap_xy = false,
+            .mirror_x = false,
+            .mirror_y = false,
+        },
+        .flags = {
+            .buff_dma = cfg->flags.buff_dma,
+            .buff_spiram = cfg->flags.buff_spiram,
+#if LVGL_VERSION_MAJOR >= 9
+            .swap_bytes = (BSP_LCD_BIGENDIAN ? true : false),
+#endif
+#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
+            .sw_rotate = false,                /* Avoid tearing is not supported for SW rotation */
+#else
+            .sw_rotate = cfg->flags.sw_rotate,
+#endif
+#if CONFIG_BSP_DISPLAY_LVGL_FULL_REFRESH
+            .full_refresh = true,
+#elif CONFIG_BSP_DISPLAY_LVGL_DIRECT_MODE
+            .direct_mode = true,
+#endif
+        }
+    };
+    const lvgl_port_display_dsi_cfg_t dpi_cfg = {
+        .flags = {
+#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
+            .avoid_tearing = true,
+#else
+            .avoid_tearing = false,
+#endif
+        }
+    };
+
+    return lvgl_port_add_disp_dsi(&disp_cfg, &dpi_cfg);
+}
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+
+esp_err_t bsp_touch_new(const bsp_touch_config_t *config, esp_lcd_touch_handle_t *ret_touch)
+{
+    /* Initilize I2C */
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_i2c_init());
+    /* Enable Feature */
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_feature_enable(BSP_FEATURE_TOUCH, true));
+
+    /* Initialize touch */
+    esp_lcd_touch_config_t tp_cfg = {
+        .x_max = BSP_LCD_H_RES,
+        .y_max = BSP_LCD_V_RES,
+        .rst_gpio_num = GPIO_NUM_NC, // Usually shared with LCD reset
+        .int_gpio_num = BSP_LCD_TOUCH_INT,
+        .levels = {
+            .reset = 0,
+            .interrupt = 0,
+        },
+        .flags = {
+            .swap_xy = false,
+            .mirror_x = false,
+            .mirror_y = false,
+        },
+    };
+
+    esp_lcd_panel_io_handle_t tp_io_handle = NULL;
+    /* Select LCD Touch by board version */
+    if (bsp_get_board_version() == 1) {
+
+        /*
+        * Keep LCD touch interrupt pin in low for working touch
+        * Note: This is the fix (ver 1) - there is resistor to 3V3 on interrupt pin
+        * which is blocking GT911 touch.
+        */
+        const gpio_config_t int_gpio_config = {
+            .mode = GPIO_MODE_OUTPUT,
+            .intr_type = GPIO_INTR_DISABLE,
+            .pull_down_en = 0,
+            .pull_up_en = 1,
+            .pin_bit_mask = BIT64(BSP_LCD_TOUCH_INT),
+        };
+        gpio_config(&int_gpio_config);
+        gpio_set_level(BSP_LCD_TOUCH_INT, 0);
+
+        tp_cfg.int_gpio_num = GPIO_NUM_NC;
+        esp_lcd_panel_io_i2c_config_t tp_io_config = ESP_LCD_TOUCH_IO_I2C_GT911_CONFIG();
+        tp_io_config.dev_addr = ESP_LCD_TOUCH_IO_I2C_GT911_ADDRESS_BACKUP;
+        ESP_RETURN_ON_ERROR(esp_lcd_new_panel_io_i2c(bsp_i2c_get_handle(), &tp_io_config, &tp_io_handle), TAG, "");
+        ESP_RETURN_ON_ERROR(esp_lcd_touch_new_i2c_gt911(tp_io_handle, &tp_cfg, ret_touch),
+                            TAG, "New touch driver initialization failed");
+    } else if (bsp_get_board_version() == 2) {
+        const esp_lcd_panel_io_i2c_config_t tp_io_config = ESP_LCD_TOUCH_IO_I2C_ST7123_CONFIG();
+        ESP_RETURN_ON_ERROR(esp_lcd_new_panel_io_i2c(bsp_i2c_get_handle(), &tp_io_config, &tp_io_handle), TAG, "");
+        ESP_RETURN_ON_ERROR(esp_lcd_touch_new_i2c_st7123(tp_io_handle, &tp_cfg, ret_touch),
+                            TAG, "New touch driver initialization failed");
+    }
+
+    return ESP_OK;
+}
+
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+static lv_indev_t *bsp_display_indev_touch_init(lv_display_t *disp)
+{
+    BSP_ERROR_CHECK_RETURN_NULL(bsp_touch_new(NULL, &tp));
+    assert(tp);
+
+    /* Add touch input (for selected screen) */
+    const lvgl_port_touch_cfg_t touch_cfg = {
+        .disp = disp,
+        .handle = tp,
+    };
+
+    return lvgl_port_add_touch(&touch_cfg);
+}
+
+static esp_err_t bsp_touch_enter_sleep(void)
+{
+    assert(tp);
+    return esp_lcd_touch_enter_sleep(tp);
+}
+
+static esp_err_t bsp_touch_exit_sleep(void)
+{
+    assert(tp);
+    return esp_lcd_touch_exit_sleep(tp);
+}
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+
+#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
+lv_display_t *bsp_display_start(void)
+{
+    bsp_display_cfg_t cfg = {
+        .lvgl_port_cfg = ESP_LVGL_PORT_INIT_CONFIG(),
+        .buffer_size = BSP_LCD_H_RES * CONFIG_BSP_LCD_DRAW_BUF_HEIGHT,
+#if CONFIG_BSP_LCD_DRAW_BUF_DOUBLE
+        .double_buffer = 1,
+#else
+        .double_buffer = 0,
+#endif
+        .flags = {
+            .buff_dma = true,
+            .buff_spiram = false,
+            .sw_rotate = true,
+        }
+    };
+    return bsp_display_start_with_config(&cfg);
+}
+
+lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg)
+{
+    assert(cfg != NULL);
+    BSP_ERROR_CHECK_RETURN_NULL(lvgl_port_init(&cfg->lvgl_port_cfg));
+    BSP_NULL_CHECK(disp = bsp_display_lcd_init(cfg), NULL);
+    BSP_NULL_CHECK(disp_indev_touch = bsp_display_indev_touch_init(disp), NULL);
+
+    return disp;
+}
+
+lv_indev_t *bsp_display_get_input_dev(void)
+{
+    return disp_indev_touch;
+}
+
+void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation)
+{
+    lv_disp_set_rotation(disp, rotation);
+}
+
+bool bsp_display_lock(uint32_t timeout_ms)
+{
+    return lvgl_port_lock(timeout_ms);
+}
+
+void bsp_display_unlock(void)
+{
+    lvgl_port_unlock();
+}
+
+static esp_err_t bsp_lcd_enter_sleep(void)
+{
+    assert(disp_handles.panel);
+    return esp_lcd_panel_disp_on_off(disp_handles.panel, false);
+}
+
+static esp_err_t bsp_lcd_exit_sleep(void)
+{
+    assert(disp_handles.panel);
+    return esp_lcd_panel_disp_on_off(disp_handles.panel, true);
+}
+
+esp_err_t bsp_display_enter_sleep(void)
+{
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_lcd_enter_sleep());
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_display_backlight_off());
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_touch_enter_sleep());
+    return ESP_OK;
+}
+
+esp_err_t bsp_display_exit_sleep(void)
+{
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_lcd_exit_sleep());
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_display_backlight_on());
+    BSP_ERROR_CHECK_RETURN_ERR(bsp_touch_exit_sleep());
+    return ESP_OK;
+}
+#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
