+++ bsp/m5stack_tab5/src/bsp_feature_en.c
@@ -1,0 +1,64 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "esp_err.h"
+#include "esp_check.h"
+#include "driver/gpio.h"
+
+#include "bsp/m5stack_tab5.h"
+
+/* Feature enable */
+esp_err_t bsp_feature_enable(bsp_feature_t feature, bool enable)
+{
+    esp_err_t ret = ESP_OK;
+
+    switch (feature) {
+    case BSP_FEATURE_LCD: {
+        esp_io_expander_handle_t io_expander = bsp_io_expander_init();
+        ret |= esp_io_expander_set_dir(io_expander, BSP_LCD_EN, IO_EXPANDER_OUTPUT);
+        ret |= esp_io_expander_set_level(io_expander, BSP_LCD_EN, enable);
+        ret |= esp_io_expander_set_output_mode(io_expander, BSP_LCD_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
+        break;
+    }
+    case BSP_FEATURE_TOUCH: {
+        esp_io_expander_handle_t io_expander = bsp_io_expander_init();
+        ret |= esp_io_expander_set_dir(io_expander, BSP_TOUCH_EN, IO_EXPANDER_OUTPUT);
+        ret |= esp_io_expander_set_level(io_expander, BSP_TOUCH_EN, enable);
+        ret |= esp_io_expander_set_output_mode(io_expander, BSP_TOUCH_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
+        break;
+    }
+    case BSP_FEATURE_SPEAKER: {
+        esp_io_expander_handle_t io_expander = bsp_io_expander_init();
+        ret |= esp_io_expander_set_dir(io_expander, BSP_SPEAKER_EN, IO_EXPANDER_OUTPUT);
+        ret |= esp_io_expander_set_level(io_expander, BSP_SPEAKER_EN, enable);
+        ret |= esp_io_expander_set_output_mode(io_expander, BSP_SPEAKER_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
+        break;
+    }
+    case BSP_FEATURE_CAMERA: {
+        esp_io_expander_handle_t io_expander = bsp_io_expander_init();
+        ret |= esp_io_expander_set_dir(io_expander, BSP_CAMERA_EN, IO_EXPANDER_OUTPUT);
+        ret |= esp_io_expander_set_level(io_expander, BSP_CAMERA_EN, enable);
+        ret |= esp_io_expander_set_output_mode(io_expander, BSP_CAMERA_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
+        break;
+    }
+    case BSP_FEATURE_USB: {
+        esp_io_expander_handle_t io_expander = bsp_io_expander1_init();
+        ret |= esp_io_expander_set_dir(io_expander, BSP_USB_EN, IO_EXPANDER_OUTPUT);
+        ret |= esp_io_expander_set_level(io_expander, BSP_USB_EN, enable);
+        ret |= esp_io_expander_set_output_mode(io_expander, BSP_USB_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
+        break;
+    }
+    case BSP_FEATURE_WIFI: {
+        esp_io_expander_handle_t io_expander = bsp_io_expander1_init();
+        ret |= esp_io_expander_set_dir(io_expander, BSP_WIFI_EN, IO_EXPANDER_OUTPUT);
+        ret |= esp_io_expander_set_level(io_expander, BSP_WIFI_EN, enable);
+        ret |= esp_io_expander_set_output_mode(io_expander, BSP_WIFI_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
+        break;
+    }
+    }
+
+    return ret;
+}
