--- bsp/m5stack_tab5/src/m5stack_tab5.c
@@ -1,1051 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2025-2026 Espressif Systems (Shanghai) CO LTD
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#include <string.h>
-#include "driver/gpio.h"
-#include "driver/ledc.h"
-#include "esp_err.h"
-#include "esp_log.h"
-#include "esp_check.h"
-#include "esp_spiffs.h"
-#include "esp_lcd_panel_io.h"
-#include "esp_lcd_panel_vendor.h"
-#include "esp_lcd_panel_ops.h"
-#include "esp_lcd_mipi_dsi.h"
-#include "esp_ldo_regulator.h"
-#include "sd_pwr_ctrl_by_on_chip_ldo.h"
-#include "esp_lcd_ili9881c.h"
-#include "disp_init_data.h"
-#include "esp_io_expander_pi4ioe5v6408.h"
-#include "esp_lcd_touch_gt911.h"
-#include "esp_vfs_fat.h"
-#include "sd_pwr_ctrl_by_on_chip_ldo.h"
-#include "bsp/m5stack_tab5.h"
-#include "bsp/display.h"
-#include "bsp/touch.h"
-#include "bsp_err_check.h"
-#include "esp_codec_dev_defaults.h"
-#include "esp_video_device.h"
-#include "esp_video_init.h"
-#include "usb/usb_host.h"
-
-static const char *TAG = "m5stack_tab5";
-
-/* Can be used for i2s_std_gpio_config_t and/or i2s_std_config_t initialization */
-#define BSP_I2S_GPIO_CFG       \
-    {                          \
-        .mclk = BSP_I2S_MCLK,  \
-        .bclk = BSP_I2S_SCLK,  \
-        .ws = BSP_I2S_LCLK,    \
-        .dout = BSP_I2S_DOUT,  \
-        .din = BSP_I2S_DSIN,   \
-        .invert_flags = {      \
-            .mclk_inv = false, \
-            .bclk_inv = false, \
-            .ws_inv = false,   \
-        },                     \
-    }
-
-/* This configuration is used by default in bsp_audio_init() */
-#define BSP_I2S_DUPLEX_CFG(_sample_rate)                                                         \
-    {                                                                                                 \
-        .clk_cfg = I2S_STD_CLK_DEFAULT_CONFIG(_sample_rate),                                          \
-        .slot_cfg = I2S_STD_PHILIP_SLOT_DEFAULT_CONFIG(I2S_DATA_BIT_WIDTH_16BIT, I2S_SLOT_MODE_MONO), \
-        .gpio_cfg = BSP_I2S_GPIO_CFG,                                                                 \
-    }
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_display_t *disp;
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static bsp_lcd_handles_t disp_handles;
-static esp_ldo_channel_handle_t disp_phy_pwr_chan = NULL;
-static lv_indev_t *disp_indev_touch = NULL;
-static esp_lcd_touch_handle_t tp;   // LCD touch handle
-static sd_pwr_ctrl_handle_t pwr_ctrl_handle = NULL; //SD LDO handle
-static sdmmc_card_t *bsp_sdcard = NULL;    // Global uSD card handler
-static bool spi_sd_initialized = false;
-static esp_io_expander_handle_t io_expander = NULL;  // IO Expander
-static esp_io_expander_handle_t io_expander1 = NULL;  // IO Expander
-static i2s_chan_handle_t i2s_tx_chan = NULL;
-static i2s_chan_handle_t i2s_rx_chan = NULL;
-static const audio_codec_data_if_t *i2s_data_if = NULL;  /* Codec data interface */
-static TaskHandle_t usb_host_task;  // USB Host Library task
-
-
-
-/**
- * @brief I2C handle for BSP usage
- *
- * In IDF v5.4 you can call i2c_master_get_bus_handle(BSP_I2C_NUM, i2c_master_bus_handle_t *ret_handle)
- * from #include "esp_private/i2c_platform.h" to get this handle
- *
- * For IDF 5.2 and 5.3 you must call bsp_i2c_get_handle()
- */
-static i2c_master_bus_handle_t i2c_handle = NULL;
-static bool i2c_initialized = false;
-
-esp_err_t bsp_i2c_init(void)
-{
-    /* I2C was initialized before */
-    if (i2c_initialized) {
-        return ESP_OK;
-    }
-
-    const i2c_master_bus_config_t i2c_config = {
-        .i2c_port = BSP_I2C_NUM,
-        .sda_io_num = BSP_I2C_SDA,
-        .scl_io_num = BSP_I2C_SCL,
-        .clk_source = I2C_CLK_SRC_DEFAULT,
-        .flags.enable_internal_pullup = true,
-    };
-    BSP_ERROR_CHECK_RETURN_ERR(i2c_new_master_bus(&i2c_config, &i2c_handle));
-
-    i2c_initialized = true;
-    return ESP_OK;
-}
-
-esp_err_t bsp_i2c_deinit(void)
-{
-    BSP_ERROR_CHECK_RETURN_ERR(i2c_del_master_bus(i2c_handle));
-    i2c_initialized = false;
-    return ESP_OK;
-}
-
-i2c_master_bus_handle_t bsp_i2c_get_handle(void)
-{
-    bsp_i2c_init();
-    return i2c_handle;
-}
-
-esp_err_t bsp_spiffs_mount(void)
-{
-    esp_vfs_spiffs_conf_t conf = {
-        .base_path = CONFIG_BSP_SPIFFS_MOUNT_POINT,
-        .partition_label = CONFIG_BSP_SPIFFS_PARTITION_LABEL,
-        .max_files = CONFIG_BSP_SPIFFS_MAX_FILES,
-#ifdef CONFIG_BSP_SPIFFS_FORMAT_ON_MOUNT_FAIL
-        .format_if_mount_failed = true,
-#else
-        .format_if_mount_failed = false,
-#endif
-    };
-
-    esp_err_t ret_val = esp_vfs_spiffs_register(&conf);
-
-    BSP_ERROR_CHECK_RETURN_ERR(ret_val);
-
-    size_t total = 0, used = 0;
-    ret_val = esp_spiffs_info(conf.partition_label, &total, &used);
-    if (ret_val != ESP_OK) {
-        ESP_LOGE(TAG, "Failed to get SPIFFS partition information (%s)", esp_err_to_name(ret_val));
-    } else {
-        ESP_LOGI(TAG, "Partition size: total: %d, used: %d", total, used);
-    }
-
-    return ret_val;
-}
-
-esp_err_t bsp_spiffs_unmount(void)
-{
-    return esp_vfs_spiffs_unregister(CONFIG_BSP_SPIFFS_PARTITION_LABEL);
-}
-
-sdmmc_card_t *bsp_sdcard_get_handle(void)
-{
-    return bsp_sdcard;
-}
-
-void bsp_sdcard_get_sdmmc_host(const int slot, sdmmc_host_t *config)
-{
-    assert(config);
-    sdmmc_host_t host_config = SDMMC_HOST_DEFAULT();
-    host_config.slot = slot;
-    host_config.max_freq_khz = SDMMC_FREQ_HIGHSPEED;
-
-    memcpy(config, &host_config, sizeof(sdmmc_host_t));
-}
-
-void bsp_sdcard_get_sdspi_host(const int slot, sdmmc_host_t *config)
-{
-    assert(config);
-    sdmmc_host_t host_config = SDSPI_HOST_DEFAULT();
-    host_config.slot = slot;
-
-    memcpy(config, &host_config, sizeof(sdmmc_host_t));
-}
-
-void bsp_sdcard_sdmmc_get_slot(const int slot, sdmmc_slot_config_t *config)
-{
-    assert(config);
-    memset(config, 0, sizeof(sdmmc_slot_config_t));
-    /* SD card is connected to Slot 0 pins. Slot 0 uses IO MUX, so not specifying the pins here */
-    config->cd = SDMMC_SLOT_NO_CD;
-    config->wp = SDMMC_SLOT_NO_WP;
-    config->cmd = BSP_SD_CMD;
-    config->clk = BSP_SD_CLK;
-    config->d0 = BSP_SD_D0;
-    config->d1 = BSP_SD_D1;
-    config->d2 = BSP_SD_D2;
-    config->d3 = BSP_SD_D3;
-    config->width = 4;
-    config->flags = 0;
-}
-
-void bsp_sdcard_sdspi_get_slot(const spi_host_device_t spi_host, sdspi_device_config_t *config)
-{
-    assert(config);
-    memset(config, 0, sizeof(sdspi_device_config_t));
-    config->gpio_cs   = BSP_SD_SPI_CS;
-    config->gpio_cd   = SDSPI_SLOT_NO_CD;
-    config->gpio_wp   = SDSPI_SLOT_NO_WP;
-    config->gpio_int  = GPIO_NUM_NC;
-    config->host_id = spi_host;
-#if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(5, 2, 0)
-    config->gpio_wp_polarity = SDSPI_IO_ACTIVE_LOW;
-#endif
-}
-
-esp_err_t bsp_sdcard_sdmmc_mount(bsp_sdcard_cfg_t *cfg)
-{
-    sdmmc_host_t sdhost = {0};
-    sdmmc_slot_config_t sdslot = {0};
-    const esp_vfs_fat_sdmmc_mount_config_t mount_config = {
-#ifdef CONFIG_BSP_SD_FORMAT_ON_MOUNT_FAIL
-        .format_if_mount_failed = true,
-#else
-        .format_if_mount_failed = false,
-#endif
-        .max_files = 5,
-        .allocation_unit_size = 16 * 1024
-    };
-    assert(cfg);
-
-    if (!cfg->mount) {
-        cfg->mount = &mount_config;
-    }
-
-    if (!cfg->host) {
-        bsp_sdcard_get_sdmmc_host(SDMMC_HOST_SLOT_0, &sdhost);
-        cfg->host = &sdhost;
-    }
-
-    if (!cfg->slot.sdmmc) {
-        bsp_sdcard_sdmmc_get_slot(SDMMC_HOST_SLOT_0, &sdslot);
-        cfg->slot.sdmmc = &sdslot;
-    }
-
-    sd_pwr_ctrl_ldo_config_t ldo_config = {
-        .ldo_chan_id = 4,
-    };
-    esp_err_t ret = sd_pwr_ctrl_new_on_chip_ldo(&ldo_config, &pwr_ctrl_handle);
-    if (ret != ESP_OK) {
-        ESP_LOGE(TAG, "Failed to create a new on-chip LDO power control driver");
-        return ret;
-    }
-    cfg->host->pwr_ctrl_handle = pwr_ctrl_handle;
-
-#if !CONFIG_FATFS_LONG_FILENAMES
-    ESP_LOGW(TAG, "Warning: Long filenames on SD card are disabled in menuconfig!");
-#endif
-
-    return esp_vfs_fat_sdmmc_mount(BSP_SD_MOUNT_POINT, cfg->host, cfg->slot.sdmmc, cfg->mount, &bsp_sdcard);
-}
-
-esp_err_t bsp_sdcard_sdspi_mount(bsp_sdcard_cfg_t *cfg)
-{
-    sdmmc_host_t sdhost = {0};
-    sdspi_device_config_t sdslot = {0};
-    const esp_vfs_fat_sdmmc_mount_config_t mount_config = {
-#ifdef CONFIG_BSP_SD_FORMAT_ON_MOUNT_FAIL
-        .format_if_mount_failed = true,
-#else
-        .format_if_mount_failed = false,
-#endif
-        .max_files = 5,
-        .allocation_unit_size = 16 * 1024
-    };
-    assert(cfg);
-
-    ESP_LOGD(TAG, "Initialize SPI bus");
-    const spi_bus_config_t buscfg = {
-        .sclk_io_num     = BSP_SD_SPI_CLK,
-        .mosi_io_num     = BSP_SD_SPI_MOSI,
-        .miso_io_num     = BSP_SD_SPI_MISO,
-        .quadwp_io_num   = GPIO_NUM_NC,
-        .quadhd_io_num   = GPIO_NUM_NC,
-        .max_transfer_sz = 4000,
-    };
-    ESP_RETURN_ON_ERROR(spi_bus_initialize(BSP_SDSPI_HOST, &buscfg, SPI_DMA_CH_AUTO), TAG, "SPI init failed");
-    spi_sd_initialized = true;
-
-    if (!cfg->mount) {
-        cfg->mount = &mount_config;
-    }
-
-    if (!cfg->host) {
-        bsp_sdcard_get_sdspi_host(SDMMC_HOST_SLOT_0, &sdhost);
-        cfg->host = &sdhost;
-    }
-
-    if (!cfg->slot.sdspi) {
-        bsp_sdcard_sdspi_get_slot(BSP_SDSPI_HOST, &sdslot);
-        cfg->slot.sdspi = &sdslot;
-    }
-
-    sd_pwr_ctrl_ldo_config_t ldo_config = {
-        .ldo_chan_id = 4,
-    };
-    esp_err_t ret = sd_pwr_ctrl_new_on_chip_ldo(&ldo_config, &pwr_ctrl_handle);
-    if (ret != ESP_OK) {
-        ESP_LOGE(TAG, "Failed to create a new on-chip LDO power control driver");
-        return ret;
-    }
-    cfg->host->pwr_ctrl_handle = pwr_ctrl_handle;
-
-#if !CONFIG_FATFS_LONG_FILENAMES
-    ESP_LOGW(TAG, "Warning: Long filenames on SD card are disabled in menuconfig!");
-#endif
-
-    return esp_vfs_fat_sdspi_mount(BSP_SD_MOUNT_POINT, cfg->host, cfg->slot.sdspi, cfg->mount, &bsp_sdcard);
-}
-
-esp_err_t bsp_sdcard_mount(void)
-{
-    bsp_sdcard_cfg_t cfg = {0};
-    return bsp_sdcard_sdmmc_mount(&cfg);
-}
-
-esp_err_t bsp_sdcard_unmount(void)
-{
-    esp_err_t ret = ESP_OK;
-
-    if (pwr_ctrl_handle) {
-        ret |= sd_pwr_ctrl_del_on_chip_ldo(pwr_ctrl_handle);
-        pwr_ctrl_handle = NULL;
-    }
-
-    ret |= esp_vfs_fat_sdcard_unmount(BSP_SD_MOUNT_POINT, bsp_sdcard);
-    bsp_sdcard = NULL;
-    if (spi_sd_initialized) {
-        ret |= spi_bus_free(BSP_SDSPI_HOST);
-        spi_sd_initialized = false;
-    }
-
-    return ret;
-}
-
-esp_err_t bsp_audio_init(const i2s_std_config_t *i2s_config)
-{
-    esp_err_t ret = ESP_FAIL;
-    if (i2s_tx_chan && i2s_rx_chan) {
-        /* Audio was initialized before */
-        return ESP_OK;
-    }
-
-    /* Setup I2S peripheral */
-    i2s_chan_config_t chan_cfg = I2S_CHANNEL_DEFAULT_CONFIG(CONFIG_BSP_I2S_NUM, I2S_ROLE_MASTER);
-    chan_cfg.auto_clear = true; // Auto clear the legacy data in the DMA buffer
-    BSP_ERROR_CHECK_RETURN_ERR(i2s_new_channel(&chan_cfg, &i2s_tx_chan, &i2s_rx_chan));
-
-    /* Setup I2S channels */
-    const i2s_std_config_t std_cfg_default = BSP_I2S_DUPLEX_CFG(48000);
-    const i2s_std_config_t *p_i2s_cfg = &std_cfg_default;
-    if (i2s_config != NULL) {
-        p_i2s_cfg = i2s_config;
-    }
-
-    if (i2s_tx_chan != NULL) {
-        ESP_GOTO_ON_ERROR(i2s_channel_init_std_mode(i2s_tx_chan, p_i2s_cfg), err, TAG, "I2S channel initialization failed");
-        ESP_GOTO_ON_ERROR(i2s_channel_enable(i2s_tx_chan), err, TAG, "I2S enabling failed");
-    }
-    if (i2s_rx_chan != NULL) {
-        ESP_GOTO_ON_ERROR(i2s_channel_init_std_mode(i2s_rx_chan, p_i2s_cfg), err, TAG, "I2S channel initialization failed");
-        ESP_GOTO_ON_ERROR(i2s_channel_enable(i2s_rx_chan), err, TAG, "I2S enabling failed");
-    }
-
-    audio_codec_i2s_cfg_t i2s_cfg = {
-        .port = CONFIG_BSP_I2S_NUM,
-        .rx_handle = i2s_rx_chan,
-        .tx_handle = i2s_tx_chan,
-    };
-    i2s_data_if = audio_codec_new_i2s_data(&i2s_cfg);
-    BSP_NULL_CHECK_GOTO(i2s_data_if, err);
-
-    return ESP_OK;
-
-err:
-    if (i2s_tx_chan) {
-        i2s_del_channel(i2s_tx_chan);
-    }
-    if (i2s_rx_chan) {
-        i2s_del_channel(i2s_rx_chan);
-    }
-
-    return ret;
-}
-
-const audio_codec_data_if_t *bsp_audio_get_codec_itf(void)
-{
-    return i2s_data_if;
-}
-
-esp_codec_dev_handle_t bsp_audio_codec_speaker_init(void)
-{
-    const audio_codec_data_if_t *i2s_data_if = bsp_audio_get_codec_itf();
-    if (i2s_data_if == NULL) {
-        /* Initilize I2C */
-        BSP_ERROR_CHECK_RETURN_NULL(bsp_i2c_init());
-        /* Configure I2S peripheral and Power Amplifier */
-        BSP_ERROR_CHECK_RETURN_NULL(bsp_audio_init(NULL));
-        i2s_data_if = bsp_audio_get_codec_itf();
-    }
-    assert(i2s_data_if);
-    /* Enable Feature */
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_feature_enable(BSP_FEATURE_SPEAKER, true));
-
-    const audio_codec_gpio_if_t *gpio_if = audio_codec_new_gpio();
-
-    audio_codec_i2c_cfg_t i2c_cfg = {
-        .port = BSP_I2C_NUM,
-        .addr = ES8388_CODEC_DEFAULT_ADDR,
-        .bus_handle = i2c_handle,
-    };
-    const audio_codec_ctrl_if_t *i2c_ctrl_if = audio_codec_new_i2c_ctrl(&i2c_cfg);
-    BSP_NULL_CHECK(i2c_ctrl_if, NULL);
-
-    esp_codec_dev_hw_gain_t gain = {
-        .pa_voltage = 5.0,
-        .codec_dac_voltage = 3.3,
-    };
-
-    es8388_codec_cfg_t codec_cfg = {
-        .ctrl_if = i2c_ctrl_if,
-        .gpio_if = gpio_if,
-        .codec_mode = ESP_CODEC_DEV_WORK_MODE_DAC,
-        .pa_pin = BSP_POWER_AMP_IO,
-        .pa_reverted = false,
-        .master_mode = false,
-        .hw_gain = gain,
-    };
-    const audio_codec_if_t *dev = es8388_codec_new(&codec_cfg);
-    BSP_NULL_CHECK(dev, NULL);
-
-    esp_codec_dev_cfg_t codec_dev_cfg = {
-        .dev_type = ESP_CODEC_DEV_TYPE_OUT,
-        .codec_if = dev,
-        .data_if = i2s_data_if,
-    };
-    return esp_codec_dev_new(&codec_dev_cfg);
-}
-
-esp_codec_dev_handle_t bsp_audio_codec_microphone_init(void)
-{
-    const audio_codec_data_if_t *i2s_data_if = bsp_audio_get_codec_itf();
-    if (i2s_data_if == NULL) {
-        /* Initilize I2C */
-        BSP_ERROR_CHECK_RETURN_NULL(bsp_i2c_init());
-        /* Configure I2S peripheral and Power Amplifier */
-        BSP_ERROR_CHECK_RETURN_NULL(bsp_audio_init(NULL));
-        i2s_data_if = bsp_audio_get_codec_itf();
-    }
-    assert(i2s_data_if);
-
-    audio_codec_i2c_cfg_t i2c_cfg = {
-        .port = BSP_I2C_NUM,
-        .addr = ES7210_CODEC_DEFAULT_ADDR,
-        .bus_handle = i2c_handle,
-    };
-    const audio_codec_ctrl_if_t *i2c_ctrl_if = audio_codec_new_i2c_ctrl(&i2c_cfg);
-    BSP_NULL_CHECK(i2c_ctrl_if, NULL);
-
-    es7210_codec_cfg_t codec_cfg = {
-        .ctrl_if = i2c_ctrl_if,
-    };
-    const audio_codec_if_t *dev = es7210_codec_new(&codec_cfg);
-    BSP_NULL_CHECK(dev, NULL);
-
-    esp_codec_dev_cfg_t codec_dev_cfg = {
-        .dev_type = ESP_CODEC_DEV_TYPE_IN,
-        .codec_if = dev,
-        .data_if = i2s_data_if,
-    };
-    return esp_codec_dev_new(&codec_dev_cfg);
-}
-
-// Bit number used to represent command and parameter
-#define LCD_CMD_BITS           8
-#define LCD_PARAM_BITS         8
-#define LCD_LEDC_CH            CONFIG_BSP_DISPLAY_BRIGHTNESS_LEDC_CH
-
-esp_err_t bsp_display_brightness_init(void)
-{
-    // Setup LEDC peripheral for PWM backlight control
-    const ledc_channel_config_t LCD_backlight_channel = {
-        .gpio_num = BSP_LCD_BACKLIGHT,
-        .speed_mode = LEDC_LOW_SPEED_MODE,
-        .channel = LCD_LEDC_CH,
-        .intr_type = LEDC_INTR_DISABLE,
-        .timer_sel = LEDC_TIMER_0,
-        .duty = 0,
-        .hpoint = 0
-    };
-    const ledc_timer_config_t LCD_backlight_timer = {
-        .speed_mode = LEDC_LOW_SPEED_MODE,
-        .duty_resolution = LEDC_TIMER_10_BIT,
-        .timer_num = LEDC_TIMER_0,
-        .freq_hz = 5000,
-        .clk_cfg = LEDC_AUTO_CLK
-    };
-
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_channel_config(&LCD_backlight_channel));
-
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_brightness_deinit(void)
-{
-    const ledc_timer_config_t LCD_backlight_timer = {
-        .speed_mode = LEDC_LOW_SPEED_MODE,
-        .timer_num = LEDC_TIMER_0,
-        .deconfigure = 1
-    };
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_pause(LEDC_LOW_SPEED_MODE, 1));
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_timer_config(&LCD_backlight_timer));
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_brightness_set(int brightness_percent)
-{
-    if (brightness_percent > 100) {
-        brightness_percent = 100;
-    }
-    if (brightness_percent < 0) {
-        brightness_percent = 0;
-    }
-
-    ESP_LOGI(TAG, "Setting LCD backlight: %d%%", brightness_percent);
-    uint32_t duty_cycle = (1023 * brightness_percent) / 100; // LEDC resolution set to 10bits, thus: 100% = 1023
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_set_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH, duty_cycle));
-    BSP_ERROR_CHECK_RETURN_ERR(ledc_update_duty(LEDC_LOW_SPEED_MODE, LCD_LEDC_CH));
-
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_backlight_off(void)
-{
-    return bsp_display_brightness_set(0);
-}
-
-esp_err_t bsp_display_backlight_on(void)
-{
-    return bsp_display_brightness_set(100);
-}
-static esp_err_t bsp_enable_dsi_phy_power(void)
-{
-#if BSP_MIPI_DSI_PHY_PWR_LDO_CHAN > 0
-    // Turn on the power for MIPI DSI PHY, so it can go from "No Power" state to "Shutdown" state
-    esp_ldo_channel_config_t ldo_cfg = {
-        .chan_id = BSP_MIPI_DSI_PHY_PWR_LDO_CHAN,
-        .voltage_mv = BSP_MIPI_DSI_PHY_PWR_LDO_VOLTAGE_MV,
-    };
-    ESP_RETURN_ON_ERROR(esp_ldo_acquire_channel(&ldo_cfg, &disp_phy_pwr_chan), TAG, "Acquire LDO channel for DPHY failed");
-    ESP_LOGI(TAG, "MIPI DSI PHY Powered on");
-#endif // BSP_MIPI_DSI_PHY_PWR_LDO_CHAN > 0
-
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_new(const bsp_display_config_t *config, esp_lcd_panel_handle_t *ret_panel,
-                          esp_lcd_panel_io_handle_t *ret_io)
-{
-    esp_err_t ret = ESP_OK;
-    bsp_lcd_handles_t handles;
-    ret = bsp_display_new_with_handles(config, &handles);
-
-    *ret_panel = handles.panel;
-    *ret_io = handles.io;
-
-    return ret;
-}
-
-esp_err_t bsp_display_new_with_handles(const bsp_display_config_t *config, bsp_lcd_handles_t *ret_handles)
-{
-    esp_err_t ret = ESP_OK;
-
-    /* Enable Feature */
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_feature_enable(BSP_FEATURE_LCD, true));
-    ESP_RETURN_ON_ERROR(bsp_display_brightness_init(), TAG, "Brightness init failed");
-    ESP_RETURN_ON_ERROR(bsp_enable_dsi_phy_power(), TAG, "DSI PHY power failed");
-
-    /* create MIPI DSI bus first, it will initialize the DSI PHY as well */
-    esp_lcd_dsi_bus_config_t bus_config = {
-        .bus_id = 0,
-        .num_data_lanes = BSP_LCD_MIPI_DSI_LANE_NUM,
-        .phy_clk_src = config->dsi_bus.phy_clk_src,
-        .lane_bit_rate_mbps = config->dsi_bus.lane_bit_rate_mbps,
-    };
-    ESP_RETURN_ON_ERROR(esp_lcd_new_dsi_bus(&bus_config, &ret_handles->mipi_dsi_bus), TAG, "New DSI bus init failed");
-    disp_handles.mipi_dsi_bus = ret_handles->mipi_dsi_bus;
-
-    ESP_LOGI(TAG, "Install MIPI DSI LCD control panel");
-    // we use DBI interface to send LCD commands and parameters
-    esp_lcd_dbi_io_config_t dbi_config = {
-        .virtual_channel = 0,
-        .lcd_cmd_bits = 8,   // according to the LCD spec
-        .lcd_param_bits = 8, // according to the LCD spec
-    };
-    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_io_dbi(ret_handles->mipi_dsi_bus, &dbi_config, &ret_handles->io), err, TAG,
-                      "New panel IO failed");
-    disp_handles.io = ret_handles->io;
-
-    esp_lcd_dpi_panel_config_t dpi_config = {
-        .virtual_channel    = 0,
-        .dpi_clk_src        = MIPI_DSI_DPI_CLK_SRC_DEFAULT,
-        .dpi_clock_freq_mhz = 60,
-        .in_color_format    = LCD_COLOR_FMT_RGB565,
-        .num_fbs            = 1,
-        .video_timing =
-        {
-            .h_size            = BSP_LCD_H_RES,
-            .v_size            = BSP_LCD_V_RES,
-            .hsync_back_porch  = 140,
-            .hsync_pulse_width = 40,
-            .hsync_front_porch = 40,
-            .vsync_back_porch  = 20,
-            .vsync_pulse_width = 4,
-            .vsync_front_porch = 20,
-        },
-#if CONFIG_BSP_LCD_USE_DMA2D && (ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(6, 0, 0))
-        .flags.use_dma2d = true,
-#endif
-    };
-    dpi_config.num_fbs = CONFIG_BSP_LCD_DPI_BUFFER_NUMS;
-
-    ili9881c_vendor_config_t vendor_config = {
-        .init_cmds      = disp_init_data,
-        .init_cmds_size = sizeof(disp_init_data) / sizeof(disp_init_data[0]),
-        .mipi_config = {
-            .dsi_bus = ret_handles->mipi_dsi_bus,
-            .dpi_config = &dpi_config,
-            .lane_num   = BSP_LCD_MIPI_DSI_LANE_NUM,
-        },
-    };
-    ESP_LOGD(TAG, "Install LCD driver");
-    esp_lcd_panel_dev_config_t panel_config = {
-        .reset_gpio_num = BSP_LCD_RST,
-        .flags.reset_active_high = 0,
-        .rgb_ele_order = BSP_LCD_COLOR_SPACE,
-        .bits_per_pixel = BSP_LCD_BITS_PER_PIXEL,
-        .vendor_config = &vendor_config,
-    };
-    ESP_GOTO_ON_ERROR(esp_lcd_new_panel_ili9881c(ret_handles->io, (const esp_lcd_panel_dev_config_t *)&panel_config,
-                      &ret_handles->panel), err, TAG, "New panel failed");
-    disp_handles.panel = ret_handles->panel;
-#if CONFIG_BSP_LCD_USE_DMA2D && (ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(6, 0, 0))
-    ESP_GOTO_ON_ERROR(esp_lcd_dpi_panel_enable_dma2d(ret_handles->panel), err, TAG, "LCD panel enable DMA2D failed");
-#endif
-
-    esp_lcd_panel_reset(ret_handles->panel);
-    esp_lcd_panel_init(ret_handles->panel);
-    esp_lcd_panel_invert_color(ret_handles->panel, false);
-    esp_lcd_panel_swap_xy(ret_handles->panel, false);
-    esp_lcd_panel_mirror(ret_handles->panel, false, false);
-
-    ESP_LOGI(TAG, "Display initialized with resolution %dx%d", BSP_LCD_H_RES, BSP_LCD_V_RES);
-    return ret;
-
-err:
-    bsp_display_delete();
-    return ret;
-}
-
-void bsp_display_delete(void)
-{
-    if (disp_handles.panel) {
-        esp_lcd_panel_del(disp_handles.panel);
-        disp_handles.panel = NULL;
-    }
-    if (disp_handles.io) {
-        esp_lcd_panel_io_del(disp_handles.io);
-        disp_handles.io = NULL;
-    }
-    if (disp_handles.mipi_dsi_bus) {
-        esp_lcd_del_dsi_bus(disp_handles.mipi_dsi_bus);
-        disp_handles.mipi_dsi_bus = NULL;
-    }
-
-    if (disp_phy_pwr_chan) {
-        esp_ldo_release_channel(disp_phy_pwr_chan);
-        disp_phy_pwr_chan = NULL;
-    }
-
-    bsp_display_brightness_deinit();
-}
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_display_t *bsp_display_lcd_init(const bsp_display_cfg_t *cfg)
-{
-    assert(cfg != NULL);
-    esp_lcd_panel_io_handle_t io_handle = NULL;
-    const bsp_display_config_t bsp_disp_cfg = {
-        .dsi_bus = {
-            .phy_clk_src = 0, // let the driver to choose the default clock source
-            .lane_bit_rate_mbps = BSP_LCD_MIPI_DSI_LANE_BITRATE_MBPS,
-        }
-    };
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_new(&bsp_disp_cfg, &disp_handles.panel, &io_handle));
-
-    esp_lcd_panel_disp_on_off(disp_handles.panel, true);
-
-    /* Add LCD screen */
-    ESP_LOGD(TAG, "Add LCD screen");
-    const lvgl_port_display_cfg_t disp_cfg = {
-        .io_handle = io_handle,
-        .panel_handle = disp_handles.panel,
-        .buffer_size = cfg->buffer_size,
-        .double_buffer = cfg->double_buffer,
-        .hres = BSP_LCD_H_RES,
-        .vres = BSP_LCD_V_RES,
-        .monochrome = false,
-        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
-        .rotation = {
-            .swap_xy = false,
-            .mirror_x = false,
-            .mirror_y = false,
-        },
-        .flags = {
-            .buff_dma = cfg->flags.buff_dma,
-            .buff_spiram = cfg->flags.buff_spiram,
-#if LVGL_VERSION_MAJOR >= 9
-            .swap_bytes = (BSP_LCD_BIGENDIAN ? true : false),
-#endif
-#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
-            .sw_rotate = false,                /* Avoid tearing is not supported for SW rotation */
-#else
-            .sw_rotate = cfg->flags.sw_rotate,
-#endif
-#if CONFIG_BSP_DISPLAY_LVGL_FULL_REFRESH
-            .full_refresh = true,
-#elif CONFIG_BSP_DISPLAY_LVGL_DIRECT_MODE
-            .direct_mode = true,
-#endif
-        }
-    };
-    const lvgl_port_display_dsi_cfg_t dpi_cfg = {
-        .flags = {
-#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
-            .avoid_tearing = true,
-#else
-            .avoid_tearing = false,
-#endif
-        }
-    };
-
-    return lvgl_port_add_disp_dsi(&disp_cfg, &dpi_cfg);
-}
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
-esp_err_t bsp_touch_new(const bsp_touch_config_t *config, esp_lcd_touch_handle_t *ret_touch)
-{
-    /* Initilize I2C */
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_i2c_init());
-    /* Enable Feature */
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_feature_enable(BSP_FEATURE_TOUCH, true));
-
-    /* Initialize touch */
-    esp_lcd_touch_config_t tp_cfg = {
-        .x_max = BSP_LCD_H_RES,
-        .y_max = BSP_LCD_V_RES,
-        .rst_gpio_num = GPIO_NUM_NC, // Usually shared with LCD reset
-        .int_gpio_num = GPIO_NUM_NC, //GPIO_NUM_23 cannot be used due to resistor to 3V3
-        .levels = {
-            .reset = 0,
-            .interrupt = 0,
-        },
-        .flags = {
-            .swap_xy = false,
-            .mirror_x = false,
-            .mirror_y = false,
-        },
-    };
-
-    /*
-     * Keep interrupt pin in low for working touch
-     * Note: This is the fix - there is resistor to 3V3 on interrupt pin
-     * which is blocking GT911 touch.
-     */
-    const gpio_config_t int_gpio_config = {
-        .mode = GPIO_MODE_OUTPUT,
-        .intr_type = GPIO_INTR_DISABLE,
-        .pull_down_en = 0,
-        .pull_up_en = 1,
-        .pin_bit_mask = BIT64(GPIO_NUM_23),
-    };
-    gpio_config(&int_gpio_config);
-    gpio_set_level(GPIO_NUM_23, 0);
-
-    esp_lcd_panel_io_handle_t tp_io_handle = NULL;
-    esp_lcd_panel_io_i2c_config_t tp_io_config = ESP_LCD_TOUCH_IO_I2C_GT911_CONFIG();
-    tp_io_config.dev_addr = ESP_LCD_TOUCH_IO_I2C_GT911_ADDRESS_BACKUP;
-    ESP_RETURN_ON_ERROR(esp_lcd_new_panel_io_i2c(i2c_handle, &tp_io_config, &tp_io_handle), TAG, "");
-
-    ESP_RETURN_ON_ERROR(esp_lcd_touch_new_i2c_gt911(tp_io_handle, &tp_cfg, ret_touch), TAG,
-                        "New touch driver initialization failed");
-
-    return ESP_OK;
-}
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_indev_t *bsp_display_indev_touch_init(lv_display_t *disp)
-{
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_touch_new(NULL, &tp));
-    assert(tp);
-
-    /* Add touch input (for selected screen) */
-    const lvgl_port_touch_cfg_t touch_cfg = {
-        .disp = disp,
-        .handle = tp,
-    };
-
-    return lvgl_port_add_touch(&touch_cfg);
-}
-
-static esp_err_t bsp_touch_enter_sleep(void)
-{
-    assert(tp);
-    return esp_lcd_touch_enter_sleep(tp);
-}
-
-static esp_err_t bsp_touch_exit_sleep(void)
-{
-    assert(tp);
-    return esp_lcd_touch_exit_sleep(tp);
-}
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-lv_display_t *bsp_display_start(void)
-{
-    bsp_display_cfg_t cfg = {
-        .lvgl_port_cfg = ESP_LVGL_PORT_INIT_CONFIG(),
-        .buffer_size = BSP_LCD_H_RES * CONFIG_BSP_LCD_DRAW_BUF_HEIGHT,
-#if CONFIG_BSP_LCD_DRAW_BUF_DOUBLE
-        .double_buffer = 1,
-#else
-        .double_buffer = 0,
-#endif
-        .flags = {
-            .buff_dma = true,
-            .buff_spiram = false,
-            .sw_rotate = true,
-        }
-    };
-    return bsp_display_start_with_config(&cfg);
-}
-
-lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg)
-{
-    assert(cfg != NULL);
-    BSP_ERROR_CHECK_RETURN_NULL(lvgl_port_init(&cfg->lvgl_port_cfg));
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_brightness_init());
-
-    BSP_NULL_CHECK(disp = bsp_display_lcd_init(cfg), NULL);
-    BSP_NULL_CHECK(disp_indev_touch = bsp_display_indev_touch_init(disp), NULL);
-
-    return disp;
-}
-
-lv_indev_t *bsp_display_get_input_dev(void)
-{
-    return disp_indev_touch;
-}
-
-void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation)
-{
-    lv_disp_set_rotation(disp, rotation);
-}
-
-bool bsp_display_lock(uint32_t timeout_ms)
-{
-    return lvgl_port_lock(timeout_ms);
-}
-
-void bsp_display_unlock(void)
-{
-    lvgl_port_unlock();
-}
-
-static esp_err_t bsp_lcd_enter_sleep(void)
-{
-    assert(disp_handles.panel);
-    return esp_lcd_panel_disp_on_off(disp_handles.panel, false);
-}
-
-static esp_err_t bsp_lcd_exit_sleep(void)
-{
-    assert(disp_handles.panel);
-    return esp_lcd_panel_disp_on_off(disp_handles.panel, true);
-}
-
-esp_err_t bsp_display_enter_sleep(void)
-{
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_lcd_enter_sleep());
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_display_backlight_off());
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_touch_enter_sleep());
-    return ESP_OK;
-}
-
-esp_err_t bsp_display_exit_sleep(void)
-{
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_lcd_exit_sleep());
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_display_backlight_on());
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_touch_exit_sleep());
-    return ESP_OK;
-}
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
-esp_err_t bsp_camera_start(const bsp_camera_cfg_t *cfg)
-{
-
-    /* Initilize I2C */
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_i2c_init());
-    /* Enable Feature */
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_feature_enable(BSP_FEATURE_CAMERA, true));
-    vTaskDelay(pdMS_TO_TICKS(100));
-
-    const esp_video_init_csi_config_t base_csi_config = {
-        .sccb_config = {
-            .init_sccb = false,
-            .i2c_handle = i2c_handle,
-            .freq = 400000,
-        },
-        .reset_pin = BSP_CAMERA_RST,
-        .pwdn_pin  = -1,
-    };
-
-    esp_video_init_config_t cam_config = {
-        .csi      = &base_csi_config,
-    };
-
-    return esp_video_init(&cam_config);
-}
-
-esp_io_expander_handle_t bsp_io_expander_init(void)
-{
-    if (io_expander && io_expander1) {
-        return io_expander;
-    }
-    /* Initilize I2C */
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_i2c_init());
-
-    esp_io_expander_new_i2c_pi4ioe5v6408(i2c_handle, BSP_IO_EXPANDER_ADDRESS, &io_expander);
-    esp_io_expander_new_i2c_pi4ioe5v6408(i2c_handle, BSP_IO_EXPANDER_ADDRESS_1, &io_expander1);
-
-    return io_expander;
-}
-
-/* Feature enable */
-esp_err_t bsp_feature_enable(bsp_feature_t feature, bool enable)
-{
-    esp_err_t ret = ESP_OK;
-
-    switch (feature) {
-    case BSP_FEATURE_LCD: {
-        bsp_io_expander_init();
-        ret |= esp_io_expander_set_dir(io_expander, BSP_LCD_EN, IO_EXPANDER_OUTPUT);
-        ret |= esp_io_expander_set_level(io_expander, BSP_LCD_EN, enable);
-        ret |= esp_io_expander_set_output_mode(io_expander, BSP_LCD_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
-        break;
-    }
-    case BSP_FEATURE_TOUCH: {
-        bsp_io_expander_init();
-        ret |= esp_io_expander_set_dir(io_expander, BSP_TOUCH_EN, IO_EXPANDER_OUTPUT);
-        ret |= esp_io_expander_set_level(io_expander, BSP_TOUCH_EN, enable);
-        ret |= esp_io_expander_set_output_mode(io_expander, BSP_TOUCH_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
-        break;
-    }
-    case BSP_FEATURE_SPEAKER: {
-        bsp_io_expander_init();
-        ret |= esp_io_expander_set_dir(io_expander, BSP_SPEAKER_EN, IO_EXPANDER_OUTPUT);
-        ret |= esp_io_expander_set_level(io_expander, BSP_SPEAKER_EN, enable);
-        ret |= esp_io_expander_set_output_mode(io_expander, BSP_SPEAKER_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
-        break;
-    }
-    case BSP_FEATURE_CAMERA: {
-        bsp_io_expander_init();
-        ret |= esp_io_expander_set_dir(io_expander, BSP_CAMERA_EN, IO_EXPANDER_OUTPUT);
-        ret |= esp_io_expander_set_level(io_expander, BSP_CAMERA_EN, enable);
-        ret |= esp_io_expander_set_output_mode(io_expander, BSP_CAMERA_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
-        break;
-    }
-    case BSP_FEATURE_USB: {
-        bsp_io_expander_init();
-        ret |= esp_io_expander_set_dir(io_expander1, BSP_USB_EN, IO_EXPANDER_OUTPUT);
-        ret |= esp_io_expander_set_level(io_expander1, BSP_USB_EN, enable);
-        ret |= esp_io_expander_set_output_mode(io_expander1, BSP_USB_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
-        break;
-    }
-    case BSP_FEATURE_WIFI: {
-        bsp_io_expander_init();
-        ret |= esp_io_expander_set_dir(io_expander1, BSP_WIFI_EN, IO_EXPANDER_OUTPUT);
-        ret |= esp_io_expander_set_level(io_expander1, BSP_WIFI_EN, enable);
-        ret |= esp_io_expander_set_output_mode(io_expander1, BSP_WIFI_EN, IO_EXPANDER_OUTPUT_MODE_PUSH_PULL);
-        break;
-    }
-    }
-
-    return ret;
-}
-
-static void usb_lib_task(void *arg)
-{
-    while (1) {
-        // Start handling system events
-        uint32_t event_flags;
-        usb_host_lib_handle_events(portMAX_DELAY, &event_flags);
-        if (event_flags & USB_HOST_LIB_EVENT_FLAGS_NO_CLIENTS) {
-            ESP_ERROR_CHECK(usb_host_device_free_all());
-        }
-        if (event_flags & USB_HOST_LIB_EVENT_FLAGS_ALL_FREE) {
-            ESP_LOGI(TAG, "USB: All devices freed");
-            // Continue handling USB events to allow device reconnection
-            // The only way this task can be stopped is by calling bsp_usb_host_stop()
-        }
-    }
-}
-
-esp_err_t bsp_usb_host_start(bsp_usb_host_power_mode_t mode, bool limit_500mA)
-{
-    /* Enable Feature */
-    BSP_ERROR_CHECK_RETURN_ERR(bsp_feature_enable(BSP_FEATURE_USB, true));
-
-    //Install USB Host driver. Should only be called once in entire application
-    ESP_LOGI(TAG, "Installing USB Host");
-    const usb_host_config_t host_config = {
-        .skip_phy_setup = false,
-        .intr_flags = ESP_INTR_FLAG_LOWMED,
-    };
-    BSP_ERROR_CHECK_RETURN_ERR(usb_host_install(&host_config));
-
-    // Create a task that will handle USB library events
-    if (xTaskCreate(usb_lib_task, "usb_lib", 4096, NULL, 10, &usb_host_task) != pdTRUE) {
-        ESP_LOGE(TAG, "Creating USB host lib task failed");
-        abort();
-    }
-
-    return ESP_OK;
-}
-
-esp_err_t bsp_usb_host_stop(void)
-{
-    usb_host_uninstall();
-    if (usb_host_task) {
-        vTaskSuspend(usb_host_task);
-        vTaskDelete(usb_host_task);
-    }
-    return ESP_OK;
-}
