+++ examples/display_sensors/sdkconfig.bsp.m5stack_tab5
@@ -8,15 +8,12 @@ CONFIG_ESPTOOLPY_FLASHSIZE_4MB=y
 
 CONFIG_ESP32P4_SELECTS_REV_LESS_V3=y
 CONFIG_CODEC_I2C_BACKWARD_COMPATIBLE=n
-CONFIG_CAM_CTRL_SPI_ENABLE=n
 
 
 # SPIRAM for ESP32-P4
 CONFIG_SPIRAM=y
 CONFIG_SPIRAM_SPEED_200M=y
 CONFIG_IDF_EXPERIMENTAL_FEATURES=y
-
-
 
 ## LVGL9 ##
 CONFIG_LV_CONF_SKIP=y
